Enabled lazy imports for fastai models
